discord-bot/requirements.txt
discord.py>=2.4,<3
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
